Address matching in the Webserver wasn't actually working, because casting of integers to bytes was implemented wrong. PR: XMLRPC-145 Submitted-by: Gamaliel Amaudruz, gamaliel@fastmail.fm
git-svn-id: https://svn.apache.org/repos/asf/webservices/xmlrpc/trunk@561948 13f79535-47bb-0310-9956-ffa450edef68

+++ server/src/main/java/org/apache/xmlrpc/webserver/WebServer.java
@@ -86,7 +86,10 @@ public class WebServer implements Runnable {
 					if ("*".equals(next)) {
 						pattern[i] = 256;
 					} else {
-						pattern[i] = Integer.parseInt(next);
+					    /* Note: *Not* pattern[i] = Integer.parseInt(next);
+					     * See XMLRPC-145
+					     */
+						pattern[i] = (byte) Integer.parseInt(next);
 					}
 				}
 			} catch (Exception e) {
